+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-All notable changes in keycloak-subject-id-attribute-mapper will be documented in this file.
+All notable changes in keycloak-idp-mappers will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support for User attribute Template Importer
+- Support for Hashed User ID Mapper
+- Support for EduPersonTargetedID Mapper
 
 ## [1.1.0] - 2024-10-23
 It is based on Keycloak version 22.